# app/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Swagger Gai Identity - OpenAPI 3.0
  description: |-
    Este é um documento tratasse da especificação OpenAPI 3.0 do Micro Serviço Gai Identity.
    O Gai Identity é um serviço que faz parte do ecossistema Brazona Tech, e é utilizado para gerenciar a identidade e o acesso dos usuários em diversos aplicativos e serviços.    
    
    Alguns links úteis:
    - [Repositório de versionamento](https://github.com/brazona/bzn-gai-services)
    - [Documentação da aplicação](https://github.com/brazona/bzn-gai-services)
  termsOfService: https://swagger.io/terms/
  contact:
    email: suporte@brazona.com.br
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: /bzn-gai-services-identity
    description: Servidor de Produção do Gai Identity
...
paths:
...
  /addresses/v1/{username}/{name}:
    put:
      tags:
        - addresses
      summary: Atualiza endereço de um usuário.
      description: Atualiza as endereço de um usuário existente no sistema.
      operationId: updateAddressByUsernameAndName
      parameters:
        - name: username
          in: path
          description: Nome de usuário para busca
          required: true
          schema:
            type: string
        - name: name
          in: path
          description: Nome do endereço para busca
          required: true
          schema:
            type: string
      requestBody:
        description: Objeto de endereço para atualização  
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressModel'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error" 
    
    get:
      tags:
        - addresses
      summary: Busca endereço por nome de usuário e nome do endereço.
      description: Retorna as endereço associadas ao nome de usuário fornecido e nome do endereço.
      operationId: getAddressByUsernameAndName
      parameters:
        - name: username
          in: path
          description: Nome de usuário para busca
          required: true
          schema:
            type: string 
        - name: name
          in: path
          description: Nome do endereço para busca
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressModel'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - addresses
      summary: Deleta endereço por nome de usuário e nome do endereço.
      description: Deleta o endereço associadas ao nome de usuário fornecido.
      operationId: deleteAddressByUsernameAndName
      parameters:
        - name: username
          in: path
          description: Nome de usuário para busca
          required: true
          schema:
            type: string 
        - name: name
          in: path
          description: Nome do endereço para busca
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"  
  /addresses/v1/{username}/create:
    post:
      tags:
        - addresses
      summary: Cria um novo endereço.
      description: Cria um novo endereço para uma Usuário no sistema.
      operationId: createAddressByUsername
      parameters:
        - name: username
          in: path
          description: Nome de usuário para busca
          required: true
          schema:
            type: string
      requestBody:
        description: Objeto de endereço para criação
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressModel'
        required: true  
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseModel'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        '422':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /addresses/v1/{username}/list:              
    get:
      tags:
        - addresses
      summary: Busca lista endereço por nome de usuário.
      description: Retorna as lista endereço associadas ao nome de usuário fornecido.
      operationId: getAddressListByUsername
      parameters:
        - name: username
          in: path
          description: Nome de usuário para busca
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressModel'  
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"                            
...
components:
  schemas:
...
    AddressModel:
      type: object
      required:
        - name
        - street
        - number
        - neighborhood
        - city
        - state
        - country
        - zipCode
      properties:
        name:
          type: string
          description: Nome do endereço
          example: Endereço Residencial
        street:
          type: string
          description: Rua do endereço
          example: Rua das Flores
        number:
          type: string
          description: Número do endereço
          example: '123'
        complement:
          type: string
          description: Complemento do endereço
          example: Apt 456                                             
        neighborhood:
          type: string
          description: Bairro do endereço
          example: Jardim das Rosas  
        city:
          type: string
          description: Cidade do endereço
          example: São Paulo
        state:
          type: string
          description: Estado do endereço
          example: SP
        country:
          type: string
          description: País do endereço
          example: Brasil
        zipCode:
          type: string
          description: Código postal do endereço
          example: '12345-678'
        username:
          type: string
          description: Nome de usuário associado ao endereço
          example: joaozinho@email.com      
...
    ResponseModel:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: Código de resposta da operação
          example: 200
        message:
          type: string
          description: Mensagem de resposta da operação
          example: Operação realizada com sucesso
        dateProcessed:
          type: string
          format: date-time
          description: Data e hora em que a operação foi processada
          example: '2023-10-01T12:00:00Z'  
    Error:
      type: object
      properties:
        code:
          type: string
          description: Código de erro
          example: '400'
        message:
          type: string
          description: Mensagem de erro
          example: 'Invalid input'
        dateProcessed:
          type: string
          format: date-time
          description: Data e hora em que a operação foi processada
          example: '2023-10-01T12:00:00Z'    
        details:
          type: string
          description: Detalhes adicionais sobre o erro
          example: 'O campo "document" é obrigatório.'
      required:
        - code
        - message
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            "write:pets": modify pets in your account
            "read:pets": read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header